simulation controls
