+++ README.md
@@ -6,3 +6,6 @@ orca whale behavior simulation
 
 </div>
 
+A simulation on orca whale behavior patterns, including pod dynamics, hunting
+tactics and flocking.
+
