+++ src/ai/lifespan.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::*;
+
+#[derive(Component)]
+pub struct Age(pub usize);
