+++ src/main.rs
@@ -5,6 +5,7 @@ mod app;
 mod camera;
 mod fish;
 mod orca;
+mod sim;
 mod ui;
 
 mod prelude {}
